% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\docType{data}
\name{ABSOLUTE_CHART_TYPES}
\alias{ABSOLUTE_CHART_TYPES}
\title{Chart types that use absolute internal unit (no scaling)}
\format{
An object of class \code{character} of length 6.
}
\usage{
ABSOLUTE_CHART_TYPES
}
\description{
Chart types that use absolute internal unit (no scaling)
}
\keyword{internal}
