% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{add_comments_optimized}
\alias{add_comments_optimized}
\title{Add Comments Optimized}
\usage{
add_comments_optimized(plot, data, kommentar_column, config)
}
\arguments{
\item{plot}{ggplot object}

\item{data}{Data frame}

\item{kommentar_column}{Column name for comments}

\item{config}{Plot configuration}
}
\description{
Add comments to plot with optimization
}
