% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{add_resource_path}
\alias{add_resource_path}
\title{Application Resource Path Setup (Golem-style)}
\usage{
add_resource_path(path = "www", prefix = "www")
}
\arguments{
\item{path}{Path to add as resource}

\item{prefix}{Prefix for the resource path}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Setup static resource paths following golem conventions.
}
