% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_updates.R
\name{add_ui_update_emit_functions}
\alias{add_ui_update_emit_functions}
\title{Add UI Update Emit Functions}
\usage{
add_ui_update_emit_functions(emit, app_state)
}
\arguments{
\item{emit}{Existing emit API}

\item{app_state}{App state containing events}
}
\description{
Add UI update emit functions to the existing emit API.
}
