% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_performance.R
\name{appears_numeric}
\alias{appears_numeric}
\title{Appears Numeric}
\usage{
appears_numeric(x)
}
\arguments{
\item{x}{Character vector to check}
}
\value{
Logical indicating if data appears numeric
}
\description{
Check if a character vector appears to contain numeric data
}
