% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_performance_opt.R
\name{apply_batch_ui_updates}
\alias{apply_batch_ui_updates}
\title{Apply Batch UI Updates}
\usage{
apply_batch_ui_updates(session, updates)
}
\arguments{
\item{session}{Shiny session}

\item{updates}{List of UI updates to apply}
}
\description{
Apply multiple UI updates in a single operation
}
