% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AUTO_RESTORE_ENABLED}
\alias{AUTO_RESTORE_ENABLED}
\title{Test mode konfiguration (defined in global.R)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
AUTO_RESTORE_ENABLED
}
\description{
Test mode konfiguration (defined in global.R)
}
\keyword{datasets}
