% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{AUTOSAVE_DELAYS}
\alias{AUTOSAVE_DELAYS}
\title{Auto-save debounce delays (milliseconds)}
\format{
An object of class \code{list} of length 2.
}
\usage{
AUTOSAVE_DELAYS
}
\description{
Auto-save debounce delays (milliseconds)
}
\keyword{datasets}
