% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plot_optimizations.R
\name{build_plot_optimized}
\alias{build_plot_optimized}
\title{Build Plot with Optimization}
\usage{
build_plot_optimized(
  plot_data,
  config,
  chart_title_reactive,
  y_axis_unit,
  kommentar_column
)
}
\description{
Fast plot building with cached components
}
