% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{CACHE_CONFIG}
\alias{CACHE_CONFIG}
\title{Cache Configuration}
\format{
An object of class \code{list} of length 5.
}
\usage{
CACHE_CONFIG
}
\description{
SPRINT 3: Centralized cache timeout settings
}
\keyword{datasets}
