% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_caching.R
\name{cache_result}
\alias{cache_result}
\title{Cache Result}
\usage{
cache_result(cache_key, value, timeout_seconds)
}
\arguments{
\item{cache_key}{Character string med cache key}

\item{value}{Value at cache}

\item{timeout_seconds}{Timeout i sekunder}
}
\description{
Gemmer result i cache med expiration time.
}
