% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup_cache.R
\name{cache_startup_data}
\alias{cache_startup_data}
\title{Cache startup data}
\usage{
cache_startup_data()
}
\value{
List of successfully cached artifacts
}
\description{
Cache all configured static artifacts for faster subsequent startups
}
