% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_diff.R
\name{calculate_plot_data_hash}
\alias{calculate_plot_data_hash}
\title{Calculate Plot Data Hash}
\usage{
calculate_plot_data_hash(data, config)
}
\arguments{
\item{data}{Data frame with plot data}

\item{config}{Chart configuration with column mappings}
}
\value{
Character hash of data state
}
\description{
Generates a hash of plot data to detect data changes.
More lightweight than comparing entire data frames.
}
\examples{
data <- data.frame(x = 1:10, y = rnorm(10))
config <- list(x_col = "x", y_col = "y")
hash1 <- calculate_plot_data_hash(data, config)
}
