% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_model.R
\name{chart_type_to_ui_type}
\alias{chart_type_to_ui_type}
\title{Map diagramtype til Y-akse UI-type}
\usage{
chart_type_to_ui_type(chart_type)
}
\arguments{
\item{chart_type}{qicharts2-kode eller dansk label}
}
\value{
one of {"count","percent","rate","time"}
}
\description{
Map diagramtype til Y-akse UI-type
}
