% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHART_TYPES_DA}
\alias{CHART_TYPES_DA}
\title{Standard SPC chart typer mapping (dansk til engelsk)}
\format{
An object of class \code{list} of length 9.
}
\usage{
CHART_TYPES_DA
}
\description{
Standard SPC chart typer mapping (dansk til engelsk)
}
\keyword{datasets}
