% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHART_TYPES_EN}
\alias{CHART_TYPES_EN}
\title{Tilsvarende engelske chart typer}
\format{
An object of class \code{list} of length 10.
}
\usage{
CHART_TYPES_EN
}
\description{
Tilsvarende engelske chart typer
}
\keyword{datasets}
