% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_label_placement.R
\name{clamp01}
\alias{clamp01}
\title{Clamp værdi til interval \verb{[0, 1]}}
\usage{
clamp01(x)
}
\arguments{
\item{x}{Numerisk værdi eller vektor}
}
\value{
Værdi begrænset til \verb{[0, 1]}
}
\description{
Clamp værdi til interval \verb{[0, 1]}
}
