% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_memory_management.R
\name{cleanup_reactive_values}
\alias{cleanup_reactive_values}
\title{Clean up reactive values}
\usage{
cleanup_reactive_values(values)
}
\arguments{
\item{values}{ReactiveValues object der skal cleanes}
}
\description{
Systematisk cleanup af reactive values til at frigøre memory.
Null'er store data objekter og clearer temporary state.
}
\seealso{
Other memory_management: 
\code{\link{cleanup_app_state}()},
\code{\link{cleanup_observers}()},
\code{\link{cleanup_temp_files}()},
\code{\link{reset_app_to_clean_state}()},
\code{\link{setup_session_cleanup}()},
\code{\link{start_memory_monitoring}()}
}
\concept{memory_management}
