% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{clear_performance_caches}
\alias{clear_performance_caches}
\title{Clear Performance Caches}
\usage{
clear_performance_caches()
}
\description{
Clear all performance caches to free memory
}
