% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_updates.R
\name{comprehensive_system_cleanup}
\alias{comprehensive_system_cleanup}
\title{Comprehensive System Cleanup}
\usage{
comprehensive_system_cleanup(app_state)
}
\arguments{
\item{app_state}{The centralized app state object}
}
\description{
Performs full cleanup of all token and queue systems.
Should be called periodically to maintain system health.
}
