% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils_ui_helpers.R
\name{create_card_config}
\alias{create_card_config}
\title{Opret standard kort (card) configuration}
\usage{
create_card_config(title, height = NULL, full_screen = FALSE)
}
\arguments{
\item{title}{Titel for kortet}

\item{height}{Højde af kortet (valgfri)}

\item{full_screen}{Om kortet skal være fuld skærm}
}
\value{
Liste med kort konfiguration
}
\description{
Opret standard kort (card) configuration
}
