% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_chart_module_helpers.R
\name{create_spc_results_processor}
\alias{create_spc_results_processor}
\title{SPC Results Processor}
\usage{
create_spc_results_processor()
}
\description{
Processer SPC beregninger og ekstraherer relevante metrics.
Ekstraeret fra mod_spc_chart_server.R for bedre testability.
}
