% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{CSV_SEPARATORS}
\alias{CSV_SEPARATORS}
\title{Standard CSV separators}
\format{
Named list med separator karakterer
}
\usage{
CSV_SEPARATORS
}
\description{
Liste over understøttede kolonne separatorer til CSV fil parsing.
}
\keyword{datasets}
