% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{DEBOUNCE_DELAYS}
\alias{DEBOUNCE_DELAYS}
\title{Debounce delays for reactive operations}
\format{
An object of class \code{list} of length 3.
}
\usage{
DEBOUNCE_DELAYS
}
\description{
Debounce delays for reactive operations
}
\keyword{datasets}
