% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_model.R
\name{decide_default_y_axis_ui_type}
\alias{decide_default_y_axis_ui_type}
\title{Vælg default Y-akse UI-type ud fra kontekst}
\usage{
decide_default_y_axis_ui_type(chart_type, n_present)
}
\arguments{
\item{chart_type}{qicharts2-kode for korttype (fx "run")}

\item{n_present}{Logical – om N-kolonne er valgt}
}
\value{
"percent" eller "count"
}
\description{
Særligt for run chart ønsker vi:
\itemize{
\item Hvis både tæller og nævner er valgt: default = percent
\item Hvis kun tæller: default = count
Brugeren kan altid overskrive.
}
}
