% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{DECIMAL_SEPARATORS}
\alias{DECIMAL_SEPARATORS}
\title{Standard decimal separators}
\format{
Named list med decimal separator karakterer
}
\usage{
DECIMAL_SEPARATORS
}
\description{
Liste over understøttede decimal separatorer til numerisk parsing.
}
\keyword{datasets}
