% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_ENCODING}
\alias{DEFAULT_ENCODING}
\title{Standard encoding for Windows kompatibilitet}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_ENCODING
}
\description{
ISO-8859-1 encoding bruges for at sikre kompatibilitet med Windows systemer
og danske specialtegn i CSV filer.
}
\keyword{datasets}
