% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_PORT}
\alias{DEFAULT_PORT}
\title{Standard port for development server}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_PORT
}
\description{
Port number brugt til udvikling af Shiny applikationen.
}
\keyword{datasets}
