% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\name{detect_unit_from_data}
\alias{detect_unit_from_data}
\title{Detect unit from data using clear heuristics}
\usage{
detect_unit_from_data(y_data)
}
\arguments{
\item{y_data}{Numeric vector}
}
\value{
Character unit type
}
\description{
Detect unit from data using clear heuristics
}
\keyword{internal}
