% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_auto_restore_setting}
\alias{determine_auto_restore_setting}
\title{Determine Auto Restore Setting}
\usage{
determine_auto_restore_setting(override_options)
}
\arguments{
\item{override_options}{Override options}
}
\value{
Boolean for auto restore setting
}
\description{
Environment-aware auto restore configuration.
}
