% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_debug_mode_setting}
\alias{determine_debug_mode_setting}
\title{Determine Debug Mode Setting}
\usage{
determine_debug_mode_setting(override_options)
}
\arguments{
\item{override_options}{Override options}
}
\value{
Boolean for debug mode setting
}
\description{
Environment-aware debug mode configuration.
}
