% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_environment_type}
\alias{determine_environment_type}
\title{Determine Environment Type}
\usage{
determine_environment_type(env_config)
}
\arguments{
\item{env_config}{Environment configuration}
}
\value{
String indicating environment type
}
\description{
Determine textual environment type fra boolean flags.
}
