% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_environment_type_from_context}
\alias{determine_environment_type_from_context}
\title{Determine Environment Type from Context}
\usage{
determine_environment_type_from_context(override_options = list())
}
\arguments{
\item{override_options}{Override options}
}
\value{
String indicating environment type
}
\description{
Determine environment type using available context og override options.
}
