% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_log_level}
\alias{determine_log_level}
\title{Determine Log Level}
\usage{
determine_log_level(override_options)
}
\arguments{
\item{override_options}{Override options}
}
\value{
Log level string
}
\description{
Environment-aware log level determination.
}
