% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{determine_test_mode_setting}
\alias{determine_test_mode_setting}
\title{Determine Test Mode Setting}
\usage{
determine_test_mode_setting(override_options)
}
\arguments{
\item{override_options}{Override options}
}
\value{
Boolean for test mode setting
}
\description{
Environment-aware test mode configuration.
}
