% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_plot_optimization.R
\docType{data}
\name{.plot_cache_env}
\alias{.plot_cache_env}
\title{Plot Cache Management}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.plot_cache_env
}
\description{
In-memory cache for plot-related objects
}
\keyword{datasets}
