% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{end_perf_timer}
\alias{end_perf_timer}
\title{End Performance Timer}
\usage{
end_perf_timer(start_time, operation_name)
}
\arguments{
\item{start_time}{Start time from start_perf_timer}

\item{operation_name}{Name of operation being timed}
}
\description{
End Performance Timer
}
