% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_ASPECT_RATIO_MAX}
\alias{EXPORT_ASPECT_RATIO_MAX}
\title{Maximum aspect ratio for custom export dimensions}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EXPORT_ASPECT_RATIO_MAX
}
\description{
Forhindrer alt for brede plots (width/height > 2.0).
}
\keyword{datasets}
