% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_DEPARTMENT_MAX_LENGTH}
\alias{EXPORT_DEPARTMENT_MAX_LENGTH}
\title{Maximum character length for department name}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EXPORT_DEPARTMENT_MAX_LENGTH
}
\description{
Begrænser afdeling/afsnit navn til 100 karakterer.
}
\keyword{datasets}
