% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_DESCRIPTION_MAX_LENGTH}
\alias{EXPORT_DESCRIPTION_MAX_LENGTH}
\title{Maximum character length for export description}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EXPORT_DESCRIPTION_MAX_LENGTH
}
\description{
Begrænser beskrivelse længde til 2000 karakterer for PDF metadata.
}
\keyword{datasets}
