% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_FILENAME_PREFIX}
\alias{EXPORT_FILENAME_PREFIX}
\title{Default prefix for exported filenames}
\format{
An object of class \code{character} of length 1.
}
\usage{
EXPORT_FILENAME_PREFIX
}
\description{
Alle eksporterede filer starter med denne prefix.
}
\keyword{datasets}
