% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_PDF_CONFIG}
\alias{EXPORT_PDF_CONFIG}
\title{PDF export configuration}
\format{
Named list med PDF settings
}
\usage{
EXPORT_PDF_CONFIG
}
\description{
Standard indstillinger for PDF eksport.
}
\keyword{datasets}
