% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_PNG_CONFIG}
\alias{EXPORT_PNG_CONFIG}
\title{PNG export configuration}
\format{
Named list med PNG settings
}
\usage{
EXPORT_PNG_CONFIG
}
\description{
Standard indstillinger for PNG eksport.
}
\keyword{datasets}
