% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_POWERPOINT_CONFIG}
\alias{EXPORT_POWERPOINT_CONFIG}
\title{PowerPoint export configuration}
\format{
Named list med PowerPoint settings
}
\usage{
EXPORT_POWERPOINT_CONFIG
}
\description{
Standard indstillinger for PowerPoint eksport.
}
\keyword{datasets}
