% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_export_config.R
\docType{data}
\name{EXPORT_VALIDATION_RULES}
\alias{EXPORT_VALIDATION_RULES}
\title{Export validation rules}
\format{
Named list med validation parameters
}
\usage{
EXPORT_VALIDATION_RULES
}
\description{
Regler til validering af export inputs før generering.
}
\keyword{datasets}
