% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{favicon}
\alias{favicon}
\title{Favicon Setup (Golem-style)}
\usage{
favicon(path = "www/favicon.ico")
}
\arguments{
\item{path}{Path to favicon file}
}
\value{
HTML tags for favicon
}
\description{
Setup application favicon following golem patterns.
}
