% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_anhoej_results.R
\name{filter_latest_part}
\alias{filter_latest_part}
\title{Filtrer qic_data til seneste part hvis skift er aktivt}
\usage{
filter_latest_part(qic_data, show_phases = FALSE)
}
\arguments{
\item{qic_data}{data.frame. QIC data med part kolonne}

\item{show_phases}{logical. TRUE hvis parts/skift er aktivt}
}
\value{
data.frame. Filtreret til seneste part eller hele datasæt
}
\description{
Når brugeren har aktiveret "Skift" (parts/phases), skal beregninger
kun baseres på den seneste part, ikke hele datasættet.
}
\keyword{internal}
