% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_formatting.R
\name{format_y_axis_percent}
\alias{format_y_axis_percent}
\title{Format Y-Axis for Percentage Data}
\usage{
format_y_axis_percent()
}
\value{
ggplot2 scale_y_continuous layer for percentage formatting
}
\description{
Format Y-Axis for Percentage Data
}
\keyword{internal}
