% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_caching.R
\name{generate_data_cache_key}
\alias{generate_data_cache_key}
\title{Generate Data-Based Cache Key}
\usage{
generate_data_cache_key(data, prefix = "data", include_names = FALSE)
}
\arguments{
\item{data}{Data object (data.frame, list, etc.)}

\item{prefix}{Cache key prefix for identification}

\item{include_names}{Include column/element names i cache key}
}
\value{
Character string med cache key
}
\description{
Genererer cache key baseret på data content ved hjælp af digest.
Sikrer at cache keys ændres når data ændres.
}
\examples{
key <- generate_data_cache_key(my_data, "autodetect")
key_detailed <- generate_data_cache_key(my_data, "processing", TRUE)

}
