% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qic_caching.R
\name{generate_qic_cache_key}
\alias{generate_qic_cache_key}
\title{Generate QIC Cache Key}
\usage{
generate_qic_cache_key(data, params)
}
\arguments{
\item{data}{Data used for QIC}

\item{params}{QIC parameters (chart type, x, y, etc.)}
}
\value{
Character string cache key
}
\description{
Creates unique key based on data + parameters.
}
\details{
Uses MD5 digests of data and parameters to create a unique key.
Same data + same parameters = same key.
}
\examples{
\dontrun{
# Generate cache key
params <- list(
  x = "Dato",
  y = "Værdi",
  chart = "run"
)
key <- generate_qic_cache_key(my_data, params)
}

}
