% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_ui_helpers.R
\name{get_absolute_right_style}
\alias{get_absolute_right_style}
\title{Opret CSS style for absolute positioning (logo/brand)}
\usage{
get_absolute_right_style()
}
\value{
Karakter string med CSS styles for højre side positioning
}
\description{
Opret CSS style for absolute positioning (logo/brand)
}
