% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_actual_qic_call_counter}
\alias{get_actual_qic_call_counter}
\title{Get actual QIC call counter value}
\usage{
get_actual_qic_call_counter()
}
\value{
Integer counter value
}
\description{
Returns current value of actual qicharts2::qic() calls for performance monitoring.
}
