% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{get_app_info}
\alias{get_app_info}
\title{Get Application Information (Golem-style)}
\usage{
get_app_info()
}
\value{
List with application information
}
\description{
Retrieve information about the current application configuration.
}
