% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_auto_restore_enabled}
\alias{get_auto_restore_enabled}
\title{Get auto restore enabled setting}
\usage{
get_auto_restore_enabled()
}
\value{
Boolean indicating if auto restore is enabled
}
\description{
Get auto restore enabled setting
}
