% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_column_mappings}
\alias{get_column_mappings}
\title{Get Column Mappings}
\usage{
get_column_mappings(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Named list with x_column, y_column, n_column, etc.
}
\description{
Retrieves all current column mappings.
}
