% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils_ui_helpers.R
\name{get_column_widths}
\alias{get_column_widths}
\title{Standard UI kolonne bredder}
\usage{
get_column_widths(layout = "quarter")
}
\arguments{
\item{layout}{Type af layout ("quarter", "half", "thirds", "sidebar")}
}
\value{
Numerisk vektor med kolonne bredder
}
\description{
Standard UI kolonne bredder
}
