% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_current_data}
\alias{get_current_data}
\title{Get Current Data}
\usage{
get_current_data(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Data frame or NULL
}
\description{
Safely retrieves the current data from app_state.
}
