% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_error_count}
\alias{get_error_count}
\title{Get Error Count}
\usage{
get_error_count(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Integer count
}
\description{
Retrieves the total error count for this session.
}
