% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_helpers.R
\name{get_flex_column_style}
\alias{get_flex_column_style}
\title{Opret standard flex container style}
\usage{
get_flex_column_style()
}
\value{
Karakter string med CSS styles
}
\description{
Returnerer standard CSS flex container style for kolonnelayout
}
