% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils_ui_helpers.R
\name{get_full_width_style}
\alias{get_full_width_style}
\title{Opret CSS style for fuld bredde elementer}
\usage{
get_full_width_style()
}
\value{
Karakter string med CSS styles
}
\description{
Opret CSS style for fuld bredde elementer
}
