% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_export.R
\name{get_hospital_name_for_export}
\alias{get_hospital_name_for_export}
\title{Get Hospital Name for Export}
\usage{
get_hospital_name_for_export()
}
\value{
Character. Hospital navn.
}
\description{
Henter hospital navn fra branding config til brug i exports.
Fallback til default hvis ikke tilgængeligt.
}
\examples{
\dontrun{
hospital <- get_hospital_name_for_export()
}

}
\seealso{
Other export_helpers: 
\code{\link{extract_spc_statistics}()},
\code{\link{generate_details_string}()},
\code{\link{quarto_available}()}
}
\concept{export_helpers}
