% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_label_placement.R
\name{get_label_placement_config}
\alias{get_label_placement_config}
\title{Hent hele label placement configuration}
\usage{
get_label_placement_config()
}
\value{
List med alle konfigurationsparametre
}
\description{
Returnerer en kopi af hele konfigurationsobjektet.
Nyttig for debugging eller når flere parametre skal hentes på en gang.
}
\examples{
config <- get_label_placement_config()
gap_line <- config$relative_gap_line * label_height_npc
gap_labels <- config$relative_gap_labels * label_height_npc

}
