% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lazy_loading.R
\name{get_lazy_loading_status}
\alias{get_lazy_loading_status}
\title{Gets lazy loading status
Gets current status of all lazy loaded modules}
\usage{
get_lazy_loading_status()
}
\value{
List with module loading status
}
\description{
Gets lazy loading status
Gets current status of all lazy loaded modules
}
