% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\name{get_log_level_name}
\alias{get_log_level_name}
\title{Get current log level name}
\usage{
get_log_level_name()
}
\value{
Character string of current log level
}
\description{
Returns the current log level as a string for easy checking
and debugging purposes.
}
\examples{
\dontrun{
current_level <- get_log_level_name()
cat("Current log level:", current_level)
}

}
