% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cache_generators.R
\name{get_observer_priorities_config}
\alias{get_observer_priorities_config}
\title{Get observer priorities configuration for caching}
\usage{
get_observer_priorities_config()
}
\value{
List with observer priorities
}
\description{
Extract observer priorities that can be cached
}
