% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_runtime_config.R
\name{get_override_or_default}
\alias{get_override_or_default}
\title{Get Override Value or Default}
\usage{
get_override_or_default(override_options, key, default)
}
\arguments{
\item{override_options}{Override options list}

\item{key}{Configuration key}

\item{default}{Default value}
}
\value{
Override value or default
}
\description{
Utility function to get override value fra options eller return default.
}
