% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_package_config}
\alias{get_package_config}
\title{Get configuration value from package environment}
\usage{
get_package_config(key, default = NULL)
}
\arguments{
\item{key}{Configuration key to retrieve}

\item{default}{Default value if key not found}
}
\value{
Configuration value or default
}
\description{
Get configuration value from package environment
}
