% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_package_hospital_name}
\alias{get_package_hospital_name}
\title{Get hospital name from package environment}
\usage{
get_package_hospital_name()
}
\value{
Hospital name string
}
\description{
Get hospital name from package environment
}
