% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_package_theme}
\alias{get_package_theme}
\title{Gets hospital theme from package environment}
\usage{
get_package_theme()
}
\value{
Bootstrap theme object
}
\description{
Gets hospital theme from package environment
}
