% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_plot_object}
\alias{get_plot_object}
\title{Get Plot Object}
\usage{
get_plot_object(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
ggplot object or NULL
}
\description{
Retrieves the current plot object.
}
