% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{get_plot_warnings}
\alias{get_plot_warnings}
\title{Get Plot Warnings}
\usage{
get_plot_warnings(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Character vector of warnings
}
\description{
Retrieves current plot warnings.
}
