% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_qic_call_counter}
\alias{get_qic_call_counter}
\title{Get QIC call counter value}
\usage{
get_qic_call_counter()
}
\value{
Integer counter value
}
\description{
Returns current value of the QIC call counter for performance monitoring.
}
