% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_runtime_config}
\alias{get_runtime_config}
\title{Get the complete runtime configuration}
\usage{
get_runtime_config()
}
\value{
Runtime configuration list or NULL
}
\description{
Get the complete runtime configuration
}
