% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_ui_helpers.R
\name{get_scroll_container_style}
\alias{get_scroll_container_style}
\title{Opret scrollbar container style med maksimal højde}
\usage{
get_scroll_container_style(max_height = UI_HEIGHTS$modal_content)
}
\arguments{
\item{max_height}{Maksimal højde (default bruger konstant)}
}
\value{
Karakter string med CSS styles
}
\description{
Opret scrollbar container style med maksimal højde
}
