% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup_cache.R
\name{get_startup_cache_status}
\alias{get_startup_cache_status}
\title{Get startup cache status
Get information about the current startup cache state}
\usage{
get_startup_cache_status()
}
\value{
List with cache status information
}
\description{
Get startup cache status
Get information about the current startup cache state
}
