% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_test_mode_auto_load}
\alias{get_test_mode_auto_load}
\title{Get test mode auto load setting}
\usage{
get_test_mode_auto_load()
}
\value{
Boolean indicating if test mode auto load is enabled
}
\description{
Get test mode auto load setting
}
