% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_helpers.R
\name{get_ui_height}
\alias{get_ui_height}
\title{Standard højder for forskellige UI komponenter}
\usage{
get_ui_height(component)
}
\arguments{
\item{component}{Type af komponent ("logo", "modal_content", "chart_container", etc.)}
}
\value{
Karakter string med højde
}
\description{
Standard højder for forskellige UI komponenter
}
