% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_helpers.R
\name{get_unit_label}
\alias{get_unit_label}
\title{Konverter enheds-kode til dansk label}
\usage{
get_unit_label(unit_code, unit_list)
}
\arguments{
\item{unit_code}{Character. Kode for organisatorisk enhed}

\item{unit_list}{Named list. Mapping mellem enheder og koder}
}
\value{
Character. Dansk label for enheden
}
\description{
Konverter enheds-kode til dansk label
}
\concept{spc_helpers}
