% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plot_optimizations.R
\name{get_y_axis_label_optimized}
\alias{get_y_axis_label_optimized}
\title{Get Y-Axis Label with Optimization}
\usage{
get_y_axis_label_optimized(y_axis_unit, y_column)
}
\description{
Cached Y-axis label generation
}
