% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side value (fallback)}
}
\value{
x if not NULL, otherwise y
}
\description{
Return right-hand side if left-hand side is NULL
}
