% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{increment_actual_qic_call_counter}
\alias{increment_actual_qic_call_counter}
\title{Increment actual QIC call counter}
\usage{
increment_actual_qic_call_counter()
}
\value{
New counter value
}
\description{
Increments the actual qicharts2::qic() call counter and returns new value.
}
