% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{increment_qic_call_counter}
\alias{increment_qic_call_counter}
\title{Increment QIC call counter}
\usage{
increment_qic_call_counter()
}
\value{
New counter value
}
\description{
Increments the QIC call counter and returns new value.
}
