% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_advanced_debug.R
\name{initialize_advanced_debug}
\alias{initialize_advanced_debug}
\title{Initialize Advanced Debug System}
\usage{
initialize_advanced_debug(enable_history = TRUE, max_history_entries = 1000)
}
\arguments{
\item{enable_history}{Enable global debug history storage (default: TRUE)}

\item{max_history_entries}{Maximum number of debug entries to store (default: 1000)}
}
\description{
Sets up global debug infrastructure
}
