% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_initialization.R
\name{initialize_main_app}
\alias{initialize_main_app}
\title{Initialize Main Application Components}
\usage{
initialize_main_app()
}
\value{
List with loading results
}
\description{
Load the main app server og UI functions. Disse skal loaded sidst
efter all dependencies er tilgængelige.
}
