% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dependency_injection.R
\name{inject_dependencies}
\alias{inject_dependencies}
\title{Inject dependencies into function}
\usage{
inject_dependencies(func, deps, ...)
}
\arguments{
\item{func}{Function der skal have dependencies injected}

\item{deps}{Dependency container}

\item{...}{Additional parameters til function}
}
\value{
Function result med injected dependencies
}
\description{
Utility function til at injicere dependencies ind i functions
på en clean måde. Reducerer parameter passing complexity.
}
\examples{
\dontrun{
# Inject dependencies into function
result <- inject_dependencies(
  process_data,
  deps,
  data = my_data,
  options = my_options
)
}

}
\seealso{
Other dependency_injection: 
\code{\link{create_dependency_container}()},
\code{\link{create_repository_layer}()},
\code{\link{create_services_layer}()},
\code{\link{create_utilities_layer}()},
\code{\link{create_validation_layer}()}
}
\concept{dependency_injection}
