% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\docType{data}
\name{INTERNAL_UNITS_BY_PLOTTYPE}
\alias{INTERNAL_UNITS_BY_PLOTTYPE}
\title{Internal canonical units by plot type}
\format{
An object of class \code{list} of length 2.
}
\usage{
INTERNAL_UNITS_BY_PLOTTYPE
}
\description{
Internal canonical units by plot type
}
\keyword{internal}
