% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_anhoej_rules_hidden}
\alias{is_anhoej_rules_hidden}
\title{Check if Anhoej Rules Hidden}
\usage{
is_anhoej_rules_hidden(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if Anhoej rules should be hidden in UI.
}
