% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_file_uploaded}
\alias{is_file_uploaded}
\title{Check if File was Uploaded}
\usage{
is_file_uploaded(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if a file has been uploaded this session.
}
