% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_plot_generating}
\alias{is_plot_generating}
\title{Check if Plot Generation in Progress}
\usage{
is_plot_generating(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Checks the circuit breaker flag for overlapping plot generations.
}
