% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_plot_ready}
\alias{is_plot_ready}
\title{Check if Plot is Ready}
\usage{
is_plot_ready(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if the plot is ready to display.
}
