% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{is_prod_mode}
\alias{is_prod_mode}
\title{Check if Application is in Production Mode}
\usage{
is_prod_mode()
}
\value{
Boolean indicating production mode
}
\description{
Golem-style production mode detection.
}
