% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_font_registration.R
\name{is_roboto_available}
\alias{is_roboto_available}
\title{Check if Roboto Medium Font is Available}
\usage{
is_roboto_available()
}
\value{
Logical. TRUE hvis fonten er tilgængelig, FALSE ellers.
}
\description{
Tjekker om Roboto Medium fonten er tilgængelig i systemet.
Nyttigt for diagnostik og fejlfinding.
}
\examples{
\dontrun{
if (is_roboto_available()) {
  message("Roboto Medium er tilgængelig")
}
}

}
\seealso{
Other font_registration: 
\code{\link{get_font_status}()},
\code{\link{register_roboto_font}()}
}
\concept{font_registration}
