% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_table_updating}
\alias{is_table_updating}
\title{Check if Table is Updating}
\usage{
is_table_updating(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if a table operation is in progress.
}
