% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_test_mode_enabled}
\alias{is_test_mode_enabled}
\title{Check if Test Mode Enabled}
\usage{
is_test_mode_enabled(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if test mode is enabled.
}
