% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_user_session_started}
\alias{is_user_session_started}
\title{Check if User Started Session}
\usage{
is_user_session_started(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Safely checks if the user has actively started the session.
}
