% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{is_y_axis_autoset_done}
\alias{is_y_axis_autoset_done}
\title{Check if Y-Axis Unit Auto-Set Done}
\usage{
is_y_axis_autoset_done(app_state)
}
\arguments{
\item{app_state}{Centralized app state}
}
\value{
Logical
}
\description{
Checks if the Y-axis unit has been auto-set for this data load.
}
