% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup_cache.R
\name{load_cached_startup_data}
\alias{load_cached_startup_data}
\title{Load cached startup data}
\usage{
load_cached_startup_data()
}
\value{
List of successfully loaded artifacts with their data
}
\description{
Load cached static artifacts to speed up startup
}
