% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_initialization.R
\name{load_ui_components}
\alias{load_ui_components}
\title{Load UI Components}
\usage{
load_ui_components()
}
\value{
List with loading results
}
\description{
Load all UI-related files og components.
}
