% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LOG_COMPONENTS}
\alias{LOG_COMPONENTS}
\title{Logging komponenter for organiseret fejlfinding}
\format{
An object of class \code{list} of length 9.
}
\usage{
LOG_COMPONENTS
}
\description{
Logging komponenter for organiseret fejlfinding
}
\keyword{datasets}
