% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging.R
\docType{data}
\name{LOG_LEVELS}
\alias{LOG_LEVELS}
\title{Log levels liste til SPC App logging system}
\format{
Liste med 4 elementer
}
\usage{
LOG_LEVELS
}
\description{
Definerer numeriske værdier for forskellige log-niveauer til brug i det
konfigurerede logging-system. \strong{Lavere tal betyder højere prioritet.}
}
\details{
Log levels:
\itemize{
\item DEBUG (1): Detaljeret debug-information
\item INFO  (2): Generel information
\item WARN  (3): Advarsler
\item ERROR (4): Fejlbeskeder
}
}
\keyword{datasets}
