% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LOOP_PROTECTION_DELAYS}
\alias{LOOP_PROTECTION_DELAYS}
\title{Loop protection delays for UI updates (milliseconds)}
\format{
An object of class \code{list} of length 4.
}
\usage{
LOOP_PROTECTION_DELAYS
}
\description{
Loop protection delays for UI updates (milliseconds)
}
\keyword{datasets}
