% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server_main.R
\name{main_app_server}
\alias{main_app_server}
\title{Main Server Function}
\usage{
main_app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal Shiny parameters}
}
\description{
Main Server Function
}
