% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_caching.R
\name{manage_cache_size}
\alias{manage_cache_size}
\title{Manage Cache Size}
\usage{
manage_cache_size(max_entries)
}
\arguments{
\item{max_entries}{Maximum antal entries i cache}
}
\description{
Håndterer cache size ved at fjerne gamle entries (LRU eviction).
}
