% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\docType{data}
\name{MAX_MISSING_PERCENT}
\alias{MAX_MISSING_PERCENT}
\title{Maximum missing values procent før advarsel}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_MISSING_PERCENT
}
\description{
Maximum missing values procent før advarsel
}
\keyword{datasets}
