% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\docType{data}
\name{MIN_NUMERIC_PERCENT}
\alias{MIN_NUMERIC_PERCENT}
\title{Minimum procent numeriske værdier for kolonne detection}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_NUMERIC_PERCENT
}
\description{
Minimum procent numeriske værdier for kolonne detection
}
\keyword{datasets}
