% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_SPC_ROWS}
\alias{MIN_SPC_ROWS}
\title{Minimum antal rækker for SPC analyse}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_SPC_ROWS
}
\description{
Minimum antal rækker for SPC analyse
}
\keyword{datasets}
