% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{OPERATION_TIMEOUTS}
\alias{OPERATION_TIMEOUTS}
\title{Timeout værdier for forskellige operationer (millisekunder)}
\format{
An object of class \code{list} of length 4.
}
\usage{
OPERATION_TIMEOUTS
}
\description{
Timeout værdier for forskellige operationer (millisekunder)
}
\keyword{datasets}
