% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{PERFORMANCE_THRESHOLDS}
\alias{PERFORMANCE_THRESHOLDS}
\title{Performance monitoring thresholds}
\format{
An object of class \code{list} of length 4.
}
\usage{
PERFORMANCE_THRESHOLDS
}
\description{
Performance monitoring thresholds
}
\keyword{datasets}
