% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plot_optimizations.R
\name{prepare_qic_data_optimized}
\alias{prepare_qic_data_optimized}
\title{Prepare QIC Data with Optimization}
\usage{
prepare_qic_data_optimized(
  preprocessed_data,
  chart_type,
  target_value,
  centerline_value,
  show_phases,
  skift_column,
  frys_column
)
}
\description{
Optimized preparation of data for qicharts2
}
