% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plot_optimizations.R
\name{preprocess_spc_data_optimized}
\alias{preprocess_spc_data_optimized}
\title{Preprocess SPC Data with Optimizations}
\usage{
preprocess_spc_data_optimized(data, config)
}
\arguments{
\item{data}{Raw input data}

\item{config}{Plot configuration}
}
\description{
Batch all expensive preprocessing operations
}
