% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config_objects.R
\name{print.spc_plot_config}
\alias{print.spc_plot_config}
\title{Print SPC Plot Configuration}
\usage{
\method{print}{spc_plot_config}(x, ...)
}
\arguments{
\item{x}{SPC plot configuration object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print SPC Plot Configuration
}
