% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_observer_priorities.R
\name{priority_helpers}
\alias{priority_helpers}
\alias{priority_high}
\alias{priority_medium}
\alias{priority_low}
\alias{priority_cleanup}
\title{Observer Priority Convenience Functions}
\usage{
priority_high()

priority_medium()

priority_low()

priority_cleanup()
}
\value{
Numeric priority value
}
\description{
Quick access functions for commonly used priority levels.
}
\examples{
\dontrun{
observeEvent(input$data, priority = priority_high(), { })
}
}
