% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\docType{data}
\name{PROPORTION_CHART_TYPES}
\alias{PROPORTION_CHART_TYPES}
\title{Chart types that use proportion internal unit \link{0,1}}
\format{
An object of class \code{character} of length 3.
}
\usage{
PROPORTION_CHART_TYPES
}
\description{
Chart types that use proportion internal unit \link{0,1}
}
\keyword{internal}
