% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_label_placement.R
\name{purge_grob_cache_expired}
\alias{purge_grob_cache_expired}
\title{Purge expired entries from grob height cache}
\usage{
purge_grob_cache_expired(force = FALSE)
}
\arguments{
\item{force}{Logical: If TRUE, purge all entries regardless of TTL}
}
\value{
Integer: Number of entries purged
}
\description{
Purge expired entries from grob height cache
}
\keyword{internal}
