% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{RATE_LIMITS}
\alias{RATE_LIMITS}
\title{Rate limiting thresholds for security}
\format{
An object of class \code{list} of length 3.
}
\usage{
RATE_LIMITS
}
\description{
Rate limiting thresholds for security
}
\keyword{datasets}
