% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RECOMMENDED_SPC_POINTS}
\alias{RECOMMENDED_SPC_POINTS}
\title{Anbefalet minimum antal punkter for SPC}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RECOMMENDED_SPC_POINTS
}
\description{
Anbefalet minimum antal punkter for SPC
}
\keyword{datasets}
