% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_monitoring.R
\name{reset_qic_counters}
\alias{reset_qic_counters}
\title{Reset QIC call counters}
\usage{
reset_qic_counters()
}
\description{
Resets performance counters used to track QIC calculation calls
M1: Updated to use package environment instead of .GlobalEnv
}
