% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_export_filename.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitize Filename}
\usage{
sanitize_filename(text)
}
\arguments{
\item{text}{Input text til sanitization}
}
\value{
Sanitized filename komponent
}
\description{
Renser filnavn komponent for special characters og sikrer cross-platform
kompatibilitet. Bevarer danske karakterer (æøåÆØÅ).
}
\details{
Sanitization proces:
\enumerate{
\item Trim leading/trailing whitespace
\item Replace spaces med underscores
\item Remove special characters (keep alphanumeric, Danish chars, underscore, hyphen)
\item Collapse multiple underscores til single underscore
\item Remove leading/trailing underscores
}

Allowed characters:
\itemize{
\item A-Z, a-z (Latin letters)
\item 0-9 (Numbers)
\item æøåÆØÅ (Danish characters)
\item _ (Underscore)
\item 
\itemize{
\item (Hyphen)
}
}

Removed characters:
\itemize{
\item Path separators (/ \)
\item Special characters (@#$\%^&*()!+=[]{}|;:'",<>?)
\item Unicode characters outside allowed set
}
}
\examples{
\dontrun{
sanitize_filename("København Sygehus")
# "København_Sygehus"

sanitize_filename("Test@#$Name")
# "TestName"

sanitize_filename("  Multiple   Spaces  ")
# "Multiple_Spaces"
}

}
