% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_anhoej_rules_hidden}
\alias{set_anhoej_rules_hidden}
\title{Set Anhoej Rules Hidden Flag}
\usage{
set_anhoej_rules_hidden(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the hide Anhoej rules flag.
}
