% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_autodetect_frozen}
\alias{set_autodetect_frozen}
\title{Set Auto-Detection Frozen}
\usage{
set_autodetect_frozen(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the frozen_until_next_trigger flag.
}
