% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_file_uploaded}
\alias{set_file_uploaded}
\title{Set File Uploaded Flag}
\usage{
set_file_uploaded(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the file uploaded flag.
}
