% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_original_data}
\alias{set_original_data}
\title{Set Original Data}
\usage{
set_original_data(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Data frame to set}
}
\description{
Safely sets the original (backup) data in app_state.
}
