% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_plot_object}
\alias{set_plot_object}
\title{Set Plot Object}
\usage{
set_plot_object(app_state, plot)
}
\arguments{
\item{app_state}{Centralized app state}

\item{plot}{ggplot object}
}
\description{
Safely stores the plot object.
}
