% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_plot_ready}
\alias{set_plot_ready}
\title{Set Plot Ready Flag}
\usage{
set_plot_ready(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the plot ready flag.
}
