% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_plot_warnings}
\alias{set_plot_warnings}
\title{Set Plot Warnings}
\usage{
set_plot_warnings(app_state, warnings)
}
\arguments{
\item{app_state}{Centralized app state}

\item{warnings}{Character vector of warnings}
}
\description{
Safely sets plot warnings.
}
