% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_monitoring.R
\name{set_startup_phase}
\alias{set_startup_phase}
\title{Set startup phase}
\usage{
set_startup_phase(phase)
}
\arguments{
\item{phase}{Character. Current startup phase}
}
\description{
Set startup phase
}
