% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_table_updating}
\alias{set_table_updating}
\title{Set Table Updating Flag}
\usage{
set_table_updating(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the table updating flag.
}
