% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_test_mode_enabled}
\alias{set_test_mode_enabled}
\title{Set Test Mode Enabled}
\usage{
set_test_mode_enabled(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the test mode enabled flag.
}
