% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_test_mode_startup_phase}
\alias{set_test_mode_startup_phase}
\title{Set Test Mode Startup Phase}
\usage{
set_test_mode_startup_phase(app_state, phase)
}
\arguments{
\item{app_state}{Centralized app state}

\item{phase}{Character phase name}
}
\description{
Safely sets the test mode startup phase.
}
