% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_user_session_started}
\alias{set_user_session_started}
\title{Set User Session Started Flag}
\usage{
set_user_session_started(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the user session started flag.
}
