% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{set_y_axis_autoset_done}
\alias{set_y_axis_autoset_done}
\title{Set Y-Axis Auto-Set Done Flag}
\usage{
set_y_axis_autoset_done(app_state, value)
}
\arguments{
\item{app_state}{Centralized app state}

\item{value}{Logical value}
}
\description{
Safely sets the Y-axis auto-set done flag.
}
