% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinylogs_config.R
\name{shinylogs_dashboard_server}
\alias{shinylogs_dashboard_server}
\title{Create shinylogs dashboard module server}
\usage{
shinylogs_dashboard_server(id, log_directory = "logs/")
}
\arguments{
\item{id}{Module ID}

\item{log_directory}{Directory med log filer}
}
\description{
Server-side logik for logs dashboard.
Henter og viser logging data i real-time.
}
