% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shinylogs_config.R
\name{should_enable_shinylogs}
\alias{should_enable_shinylogs}
\title{Environment variable configuration for shinylogs}
\usage{
should_enable_shinylogs()
}
\description{
Tjekker environment variable for at kontrollere shinylogs funktioner
}
