% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{should_gc}
\alias{should_gc}
\title{Should Garbage Collect}
\usage{
should_gc()
}
\value{
Logical indicating if GC should be run
}
\description{
Intelligent decision on when to run garbage collection
}
