% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPC_ALPHA_VALUES}
\alias{SPC_ALPHA_VALUES}
\title{Alpha værdier for gennemsigtighed}
\format{
An object of class \code{list} of length 5.
}
\usage{
SPC_ALPHA_VALUES
}
\description{
Alpha værdier for gennemsigtighed
}
\keyword{datasets}
