% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPC_COLORS}
\alias{SPC_COLORS}
\title{Farve palette for SPC charts}
\format{
An object of class \code{list} of length 10.
}
\usage{
SPC_COLORS
}
\description{
Farve palette for SPC charts
}
\keyword{datasets}
