% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\docType{data}
\name{SPC_COLUMN_NAMES}
\alias{SPC_COLUMN_NAMES}
\title{Standard kolonne navne for SPC analyse}
\format{
An object of class \code{list} of length 7.
}
\usage{
SPC_COLUMN_NAMES
}
\description{
Standard kolonne navne for SPC analyse
}
\keyword{datasets}
