% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\docType{data}
\name{SPC_COMMENT_CONFIG}
\alias{SPC_COMMENT_CONFIG}
\title{Kommentar behandling konfiguration
M3: Centraliseret magic numbers for kommentar længdebegrænsninger}
\format{
An object of class \code{list} of length 3.
}
\usage{
SPC_COMMENT_CONFIG
}
\description{
Kommentar behandling konfiguration
M3: Centraliseret magic numbers for kommentar længdebegrænsninger
}
\keyword{datasets}
