% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPC_LINE_TYPES}
\alias{SPC_LINE_TYPES}
\title{Linje typer for SPC charts}
\format{
An object of class \code{list} of length 4.
}
\usage{
SPC_LINE_TYPES
}
\description{
Linje typer for SPC charts
}
\keyword{datasets}
