% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SPC_LINE_WIDTHS}
\alias{SPC_LINE_WIDTHS}
\title{Standard linje bredder}
\format{
An object of class \code{list} of length 4.
}
\usage{
SPC_LINE_WIDTHS
}
\description{
Standard linje bredder
}
\keyword{datasets}
