% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{start_perf_timer}
\alias{start_perf_timer}
\title{Performance Monitoring Utilities}
\usage{
start_perf_timer(operation_name)
}
\arguments{
\item{operation_name}{Name of operation being timed}
}
\value{
Start time
}
\description{
Start Performance Timer
}
