% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{TEST_MODE_CONFIG}
\alias{TEST_MODE_CONFIG}
\title{Test Mode Configuration}
\format{
An object of class \code{list} of length 4.
}
\usage{
TEST_MODE_CONFIG
}
\description{
SPRINT 3: Extracted magic numbers for test mode operations
}
\keyword{datasets}
