% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_monitoring.R
\name{track_event}
\alias{track_event}
\title{Track event firing}
\usage{
track_event(event_name, context = "unknown")
}
\arguments{
\item{event_name}{Character. Name of the event}

\item{context}{Character. Context where event was fired}
}
\description{
Track event firing
}
