% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_performance_monitoring.R
\name{track_generateSPCPlot_call}
\alias{track_generateSPCPlot_call}
\title{Track generateSPCPlot function call}
\usage{
track_generateSPCPlot_call(context = "unknown", details = list())
}
\arguments{
\item{context}{Character. Context where generateSPCPlot was called}

\item{details}{List. Additional context information}
}
\description{
Track generateSPCPlot function call
}
