% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ui.R
\docType{data}
\name{UI_COLUMN_WIDTHS}
\alias{UI_COLUMN_WIDTHS}
\title{Standard kolonne bredder for UI}
\format{
Named list med numeriske vektorer for kolonne bredder
}
\usage{
UI_COLUMN_WIDTHS
}
\description{
Prædefinerede kolonne bredde kombinationer til bslib layout systemer.
}
\keyword{datasets}
