% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ui_config.R
\docType{data}
\name{UI_INPUT_WIDTHS}
\alias{UI_INPUT_WIDTHS}
\title{Standard UI input widths}
\format{
An object of class \code{list} of length 5.
}
\usage{
UI_INPUT_WIDTHS
}
\description{
Standard UI input widths
}
\keyword{datasets}
