% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{UI_LAYOUT_PROPORTIONS}
\alias{UI_LAYOUT_PROPORTIONS}
\title{Layout proportions for consistent UI}
\format{
An object of class \code{list} of length 5.
}
\usage{
UI_LAYOUT_PROPORTIONS
}
\description{
Layout proportions for consistent UI
}
\keyword{datasets}
