% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_helpers.R
\docType{data}
\name{UI_SPACING}
\alias{UI_SPACING}
\title{Standard spacing konstanter for UI}
\format{
An object of class \code{list} of length 4.
}
\usage{
UI_SPACING
}
\description{
Standard spacing konstanter for UI
}
\keyword{datasets}
