% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_ui.R
\docType{data}
\name{UI_STYLES}
\alias{UI_STYLES}
\title{CSS styles constants}
\format{
Named list med CSS style strings
}
\usage{
UI_STYLES
}
\description{
Genbrugelige CSS style strings til konsistent styling.
}
\keyword{datasets}
