% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{UI_UPDATE_CONFIG}
\alias{UI_UPDATE_CONFIG}
\title{UI Update Configuration}
\format{
An object of class \code{list} of length 4.
}
\usage{
UI_UPDATE_CONFIG
}
\description{
SPRINT 3: Extracted UI update timing constants
}
\keyword{datasets}
