% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_state_accessors.R
\name{update_column_mappings}
\alias{update_column_mappings}
\title{Update All Column Mappings}
\usage{
update_column_mappings(app_state, mappings)
}
\arguments{
\item{app_state}{Centralized app state}

\item{mappings}{Named list of column mappings}
}
\description{
Safely updates multiple column mappings at once.
}
