% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{UTF8_ENCODING}
\alias{UTF8_ENCODING}
\title{Alternative encoding for UTF-8 filer}
\format{
An object of class \code{character} of length 1.
}
\usage{
UTF8_ENCODING
}
\description{
UTF-8 encoding til moderne tekst filer og internationale tegnsæt.
}
\keyword{datasets}
