% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_scaling.R
\docType{data}
\name{VALID_UNITS}
\alias{VALID_UNITS}
\title{Valid unit types for Y-axis scaling}
\format{
An object of class \code{character} of length 4.
}
\usage{
VALID_UNITS
}
\description{
Valid unit types for Y-axis scaling
}
\keyword{internal}
