% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_calculations.R
\name{validate_spc_inputs}
\alias{validate_spc_inputs}
\title{Validate input parameters for SPC plot generation}
\usage{
validate_spc_inputs(data, config)
}
\arguments{
\item{data}{Input data frame}

\item{config}{Chart configuration}
}
\description{
Validate input parameters for SPC plot generation
}
