% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_dependencies.R
\name{verify_critical_dependencies}
\alias{verify_critical_dependencies}
\title{Verify Critical Dependencies}
\usage{
verify_critical_dependencies()
}
\value{
Boolean indicating if all critical dependencies are satisfied
}
\description{
Post-loading verification that all critical functions are available.
}
