% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spc_chart.R
\name{visualizationModuleUI}
\alias{visualizationModuleUI}
\title{Visualization Module UI - Kun plot området}
\usage{
visualizationModuleUI(id)
}
\arguments{
\item{id}{Character. Namespace ID for modulet}
}
\value{
Shiny UI element
}
\description{
Håndterer selve plot visningen uden status elementer.
Status elementer er separeret i visualizationStatusUI for modularitet.
}
\seealso{
Other visualization_modules: 
\code{\link{visualizationStatusUI}()}
}
\concept{visualization_modules}
