% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_spc_config.R
\docType{data}
\name{Y_AXIS_UI_TYPES_DA}
\alias{Y_AXIS_UI_TYPES_DA}
\title{UI-typer for Y-akse (simpel valg)}
\format{
An object of class \code{list} of length 4.
}
\usage{
Y_AXIS_UI_TYPES_DA
}
\description{
UI-typer for Y-akse (simpel valg)
}
\keyword{datasets}
