% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_performance.R
\name{add_comments_optimized}
\alias{add_comments_optimized}
\title{Add Comments Optimized}
\usage{
add_comments_optimized(plot, data, kommentar_column, qic_data)
}
\arguments{
\item{plot}{ggplot object}

\item{data}{Original data frame}

\item{kommentar_column}{Column name for comments}

\item{qic_data}{QIC processed data with transformed coordinates}
}
\description{
Add comments to plot with optimization - matches extract_comment_data approach
}
