% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_performance_helpers.R
\name{appears_date}
\alias{appears_date}
\title{Appears Date}
\usage{
appears_date(x)
}
\arguments{
\item{x}{Character vector to check}
}
\value{
Logical indicating if data appears to be dates
}
\description{
Check if a character vector appears to contain date data
}
