% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system_config.R
\docType{data}
\name{AUTO_RESTORE_ENABLED}
\alias{AUTO_RESTORE_ENABLED}
\title{Auto restore funktion}
\format{
An object of class \code{logical} of length 1.
}
\usage{
AUTO_RESTORE_ENABLED
}
\description{
Bestemmer om session data automatisk skal gendannes ved app start.
}
\keyword{datasets}
