% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup_cache.R
\name{cleanup_old_cache}
\alias{cleanup_old_cache}
\title{Clean up old cache files}
\usage{
cleanup_old_cache()
}
\description{
Remove cache files that are older than their TTL or if cache size exceeds limit
}
