% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_performance_opt.R
\docType{data}
\name{.cache_env}
\alias{.cache_env}
\title{Simple Cache Implementation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.cache_env
}
\description{
In-memory cache for auto-detection results
}
\keyword{datasets}
