% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{get_golem_config_summary}
\alias{get_golem_config_summary}
\title{Get Current Golem Configuration Summary}
\usage{
get_golem_config_summary()
}
\value{
Character vector with configuration summary
}
\description{
Get human-readable summary af current golem configuration.
}
