% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spc_calculations.R
\name{process_chart_title}
\alias{process_chart_title}
\title{Process chart title for SPC plot}
\usage{
process_chart_title(chart_title_reactive, config)
}
\arguments{
\item{chart_title_reactive}{Reactive function for chart title}

\item{config}{Chart configuration with y_col}
}
\value{
Character string with processed title
}
\description{
Process chart title for SPC plot
}
