% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the SPC Shiny Application}
\usage{
run_app(port = NULL, launch_browser = NULL, ...)
}
\arguments{
\item{port}{Port number for the application (default: 3838)}

\item{launch_browser}{TRUE for browser, FALSE for no launch, or leave default for RStudio Viewer}

\item{...}{Additional arguments to pass to shinyApp()}
}
\description{
Run the SPC Shiny Application
}
