% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_autodetect_helpers.R
\name{score_by_statistical_properties}
\alias{score_by_statistical_properties}
\title{Score Statistical Properties}
\usage{
score_by_statistical_properties(col_data, role)
}
\arguments{
\item{col_data}{Vector of column data}

\item{role}{Target role ("y_column" or "n_column")}
}
\value{
Score between 0 and 1
}
\description{
Scores columns based on statistical properties relevant to SPC analysis.
}
