% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server_performance.R
\name{validate_spc_requirements}
\alias{validate_spc_requirements}
\title{Validate SPC Requirements}
\usage{
validate_spc_requirements(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\value{
Validated data frame or error
}
\description{
Validate that data meets basic SPC requirements
}
