% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{Y_AXIS_UNITS_DA}
\alias{Y_AXIS_UNITS_DA}
\title{Y-aksen enheder for forskellige chart typer}
\format{
An object of class \code{list} of length 6.
}
\usage{
Y_AXIS_UNITS_DA
}
\description{
Y-aksen enheder for forskellige chart typer
}
\keyword{datasets}
